Compiled on Linux, added the test for swapchain creation. Integrated GLEW to GlRenderer.

+++ source/RenderLib/Viewport.cpp
@@ -6,8 +6,8 @@ namespace render
 {
 	Viewport::Viewport( utils::IVec2 const & size )noexcept
 		: m_size{ size }
-		, m_viewport{ size.x, size.y, 0, 0 }
-		, m_scissor{ 0, 0, size.x, size.y }
+		, m_viewport{ uint32_t( size.x ), uint32_t( size.y ), 0, 0 }
+		, m_scissor{ 0, 0, uint32_t( size.x ), uint32_t( size.y ) }
 	{
 	}
